Major revision to file structure hopefully allows heroku to find pom.xml
