+++ src/main/java/com/lambdaschool/tiemendo/config/WebConfig.java
@@ -1,0 +1,49 @@
+//package com.lambdaschool.tiemendo.config;
+//
+//import org.apache.catalina.Context;
+//import org.apache.catalina.connector.Connector;
+//import org.apache.tomcat.util.descriptor.web.SecurityCollection;
+//import org.apache.tomcat.util.descriptor.web.SecurityConstraint;
+//import org.springframework.beans.factory.annotation.Value;
+//import org.springframework.boot.web.embedded.tomcat.TomcatServletWebServerFactory;
+//import org.springframework.boot.web.servlet.server.ServletWebServerFactory;
+//import org.springframework.context.annotation.Bean;
+//import org.springframework.context.annotation.Configuration;
+//
+//@Configuration // for HTTPS
+//public class WebConfig
+//{
+//    // This section is to add HTTP to HTTPS redirect
+//    @Value("${server.port}")
+//    private int myServerPort;
+//
+//    @Bean
+//    public ServletWebServerFactory servletContainer()
+//    {
+//        TomcatServletWebServerFactory tomcat = new TomcatServletWebServerFactory()
+//        {
+//            @Override
+//            protected void postProcessContext(Context context)
+//            {
+//                SecurityConstraint securityConstraint = new SecurityConstraint();
+//                securityConstraint.setUserConstraint("CONFIDENTIAL");
+//                SecurityCollection collection = new SecurityCollection();
+//                collection.addPattern("/*");
+//                securityConstraint.addCollection(collection);
+//                context.addConstraint(securityConstraint);
+//            }
+//        };
+//        tomcat.addAdditionalTomcatConnectors(redirectConnector());
+//        return tomcat;
+//    }
+//
+//    private Connector redirectConnector()
+//    {
+//        Connector connector = new Connector("org.apache.coyote.http11.Http11NioProtocol");
+//        connector.setScheme("http");
+//        connector.setPort(80);
+//        connector.setSecure(false);
+//        connector.setRedirectPort(myServerPort);
+//        return connector;
+//    }
+//}
