+++ src/main/java/com/lambdaschool/tiemendo/model/Role.java
@@ -1,0 +1,68 @@
+package com.lambdaschool.tiemendo.model;
+
+import com.fasterxml.jackson.annotation.JsonIgnoreProperties;
+import org.hibernate.annotations.Cascade;
+import org.hibernate.annotations.CascadeType;
+
+import javax.persistence.*;
+import java.util.ArrayList;
+import java.util.List;
+
+@Entity
+@Table(name = "roles")
+public class Role extends Auditable
+{
+
+
+    @Id
+    @GeneratedValue(strategy = GenerationType.AUTO)
+    private long roleid;
+
+    @Column(nullable = false,
+            unique = true)
+    String name;
+
+    @OneToMany(mappedBy = "role")
+    @Cascade(CascadeType.ALL)
+    @JsonIgnoreProperties("role")
+    private List<UserRoles> userRoles = new ArrayList<>();
+
+    public Role()
+    {
+    }
+
+    public Role(String name)
+    {
+        this.name = name;
+    }
+
+    public long getRoleid()
+    {
+        return roleid;
+    }
+
+    public void setRoleid(long roleid)
+    {
+        this.roleid = roleid;
+    }
+
+    public String getName()
+    {
+        return name;
+    }
+
+    public void setName(String name)
+    {
+        this.name = name;
+    }
+
+    public List<UserRoles> getUserRoles()
+    {
+        return userRoles;
+    }
+
+    public void setUserRoles(List<UserRoles> userRoles)
+    {
+        this.userRoles = userRoles;
+    }
+}
