+++ src/main/java/com/lambdaschool/tiemendo/repository/RoleRepository.java
@@ -1,0 +1,23 @@
+package com.lambdaschool.tiemendo.repository;
+
+import com.lambdaschool.tiemendo.model.Role;
+import org.springframework.data.jpa.repository.Modifying;
+import org.springframework.data.jpa.repository.Query;
+import org.springframework.data.repository.CrudRepository;
+import org.springframework.transaction.annotation.Transactional;
+
+public interface RoleRepository extends CrudRepository<Role, Long>
+{
+    @Transactional
+    @Modifying
+    @Query(value = "DELETE from UserRoles where userid = :userid")
+    void deleteUserRolesByUserId(long userid);
+
+    @Transactional
+    @Modifying
+    @Query(value = "INSERT INTO UserRoles(userid, roleid) values (:userid, :roleid)", nativeQuery = true)
+    void insertUserRoles(long userid, long roleid);
+
+    Role findRoleByName(String name);
+
+}
