+++ src/main/resources/application.properties
@@ -19,3 +19,10 @@ spring.datasource.initialization-mode=always
 spring.datasource.url=${SPRING_DATASOURCE_URL:jdbc:postgresql://localhost:5432/tiemendo}
 spring.datasource.username=${SPRING_DATASOURCE_USERNAME:tiemendo}
 spring.datasource.password=${SPRING_DATASOURCE_PASSWORD:tiemendo}
+
+spring.datasource.driverClassName=org.postgresql.Driver
+spring.datasource.maxActive=10
+spring.datasource.maxIdle=5
+spring.datasource.minIdle=2
+spring.datasource.initialSize=5
+spring.datasource.removeAbandoned=true
