--- tiemndo/src/main/java/com/lambdaschool/tiemendo/service/UserAuditing.java
@@ -1,34 +1,0 @@
-package com.lambdaschool.tiemendo.service;
-
-import org.springframework.data.domain.AuditorAware;
-import org.springframework.security.core.Authentication;
-import org.springframework.security.core.context.SecurityContextHolder;
-import org.springframework.stereotype.Component;
-
-import java.util.Optional;
-
-@Component
-public class UserAuditing implements AuditorAware<String>
-{
-
-    @Override
-    public Optional<String> getCurrentAuditor()
-    {
-//        return Optional.of("SYSTEM");
-
-
-        // for OAuth2 systems
-        String uname;
-        Authentication authentication = SecurityContextHolder.getContext().getAuthentication();
-        if (authentication != null)
-        {
-            uname = authentication.getName();
-        }
-        else
-        {
-            uname = "SYSTEM";
-        }
-        return Optional.of(uname);
-    }
-
-}
